rename: model로 패키지명 변경
