+++ src/main/java/com/shy_polarbear/server/domain/comment/model/Comment.java
@@ -1,7 +1,7 @@
-package com.shy_polarbear.server.domain.comment.entity;
+package com.shy_polarbear.server.domain.comment.model;
 
-import com.shy_polarbear.server.domain.feed.entity.Feed;
-import com.shy_polarbear.server.domain.user.entity.User;
+import com.shy_polarbear.server.domain.feed.model.Feed;
+import com.shy_polarbear.server.domain.user.model.User;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
@@ -42,6 +42,7 @@ public class Comment {
         this.author = author;
         this.content = content;
         this.feed = feed;
+        feed.getComments().add(this);
     }
 
     public static Comment createComment(User author, String content, Feed feed) {
