+++ src/main/java/com/shy_polarbear/server/domain/notification/model/Notification.java
@@ -1,7 +1,7 @@
-package com.shy_polarbear.server.domain.notification.entity;
+package com.shy_polarbear.server.domain.notification.model;
 
-import com.shy_polarbear.server.domain.comment.entity.Comment;
-import com.shy_polarbear.server.domain.user.entity.User;
+import com.shy_polarbear.server.domain.comment.model.Comment;
+import com.shy_polarbear.server.domain.user.model.User;
 import com.shy_polarbear.server.global.common.BaseEntity;
 import lombok.AccessLevel;
 import lombok.Builder;
