+++ src/main/java/com/shy_polarbear/server/domain/notification/model/NotificationType.java
@@ -1,4 +1,4 @@
-package com.shy_polarbear.server.domain.notification.entity;
+package com.shy_polarbear.server.domain.notification.model;
 
 public enum NotificationType {
     COMMENT, CHILD_COMMENT, LIMIT
