+++ src/main/java/com/shy_polarbear/server/domain/point/model/PointType.java
@@ -1,4 +1,4 @@
-package com.shy_polarbear.server.domain.point.entity;
+package com.shy_polarbear.server.domain.point.model;
 
 import lombok.Getter;
