--- src/main/java/com/shy_polarbear/server/domain/prize/entity/PrizeStatus.java
@@ -1,6 +1,0 @@
-package com.shy_polarbear.server.domain.prize.entity;
-
-public enum PrizeStatus {
-
-    GIVEN, NOT_GIVEN
-}
