+++ src/main/java/com/shy_polarbear/server/domain/prize/model/PrizeStatus.java
@@ -1,0 +1,6 @@
+package com.shy_polarbear.server.domain.prize.model;
+
+public enum PrizeStatus {
+
+    GIVEN, NOT_GIVEN
+}
