--- src/main/java/com/shy_polarbear/server/domain/user/entity/UserRole.java
@@ -1,7 +1,0 @@
-package com.shy_polarbear.server.domain.user.entity;
-
-
-
-public enum UserRole {
-    USR, ADMIN
-}
