+++ src/main/java/com/shy_polarbear/server/domain/user/model/UserRole.java
@@ -1,0 +1,7 @@
+package com.shy_polarbear.server.domain.user.model;
+
+
+
+public enum UserRole {
+    USR, ADMIN
+}
