+++ src/main/java/com/shy_polarbear/server/domain/user/model/UserStatus.java
@@ -1,4 +1,4 @@
-package com.shy_polarbear.server.domain.user.entity;
+package com.shy_polarbear.server.domain.user.model;
 
 public enum UserStatus {
     // 활성, 휴면, 탈퇴, 정지, 신고당함
