+++ src/test/java/com/shy_polarbear/server/domain/user/model/UserTest.java
@@ -1,4 +1,4 @@
-package com.shy_polarbear.server.domain.user.entity;
+package com.shy_polarbear.server.domain.user.model;
 
 
 import org.junit.jupiter.api.BeforeEach;
